move service and component to addon/
