+++ app/services/loading-slider.js
@@ -1,17 +1,1 @@
-import Ember from 'ember';
-
-const { Service, Evented } = Ember;
-
-export default Service.extend(Evented, {
-  startLoading() {
-    this.trigger('startLoading');
-  },
-
-  endLoading() {
-    this.trigger('endLoading');
-  },
-
-  changeAttrs(attrs) {
-    this.trigger('changeAttrs', attrs);
-  }
-});
+export { default } from 'ember-cli-loading-slider/services/loading-slider';
